feat(discounts): migrate discount validation logic to discount.controller and remove discountUser.controller
